HBX-1633: Move class 'org.hibernate.tool.hbm2x.XMLPrettyPrinter' to 'org.hibernate.tool.internal.xml.XMLPrettyPrinter'
